pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ros_introspect"
description = "Pure Python library for working with the files in ROS packages"
readme = "README.md"
authors = [
  { name = "David V. Lu!!", email = "davidvlu@gmail.com" },
]
maintainers = [
  { name = "David V. Lu!!", email = "davidvlu@gmail.com" },
]

dependencies = [
  "ruamel.yaml",
  "stylish-cmake-parser",
]

requires-python = ">=3.8"

dynamic = ["version"]

classifiers = [
  "Development Status :: 4 - Beta",
  "License :: OSI Approved :: BSD License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.8",
  "Topic :: Software Development",
]

[project.urls]
Homepage = "https://github.com/MetroRobots/ros_introspect"
"Bug Tracker" = "https://github.com/MetroRobots/ros_introspect/issues"

[project.optional-dependencies]
test = [
  "pytest",
  "pytest-cov",
]

[tool.hatch]
version.source = "vcs"
build.hooks.vcs.version-file = "src/ros_introspect/_version.py"
